+++ test/pipelineTest.cc
@@ -234,3 +234,31 @@ TEST(TestPipeline, bindExecutor) {
   async_simple::coro::syncAwait(std::move(pipeline).Run().via(&ex1));
   EXPECT_EQ(trans_ex, &ex2);
 }
+
+class NeverStopSource : public tunnel::Source<int> {
+ public:
+  virtual async_simple::coro::Lazy<std::optional<int>> generate() override {
+    num = num + 1;
+    co_return num;
+  }
+
+ private:
+  int num = 0;
+};
+
+TEST(TestPipeline, share_abort_channel) {
+  async_simple::executors::SimpleExecutor ex(2);
+  tunnel::Channel<int> abort_channel(10);
+  Pipeline<int> p1{PipelineOption{.bind_abort_channel = true}};
+  auto s1 = p1.AddSource(std::make_unique<SourceTest>());
+  p1.AddTransform(s1, std::make_unique<TransformTest>());
+  p1.SetSink(std::make_unique<ThrowSinkTest>());
+  Pipeline<int> p2{PipelineOption{.bind_abort_channel = true}};
+  p2.AddSource(std::make_unique<NeverStopSource>());
+  p2.SetSink(std::make_unique<SinkTest>());
+  p1.SetAbortChannel(abort_channel);
+  p2.SetAbortChannel(abort_channel);
+  std::move(p1).Run().via(&ex).start([](auto&&) {});
+  // 由于中止信息缓存在channel中，因此p1可以还没开始运行、可以正在抛出异常、甚至可以已经终止了运行，中止信息依然会传递给p2
+  async_simple::coro::syncAwait(std::move(p2).Run().via(&ex));
+}